requirements.txt
alembic==1.5.4
black==20.8b1
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-PageDown==0.3.0
Flask-SQLAlchemy==2.4.4
bootstrap-flask==1.5.1
WTForms==2.3.3
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
packaging==20.9
pathspec==0.8.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
SQLAlchemy==1.3.23